# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreatePermissionRequest:
    properties:
      action:
        description: 操作类型
        example: create
        type: string
      description:
        description: 权限描述
        example: 允许创建用户
        type: string
      display_name:
        description: 显示名称
        example: 创建用户
        type: string
      name:
        description: 权限名称
        example: user:create
        type: string
      resource:
        description: 资源类型
        example: user
        type: string
    required:
    - action
    - display_name
    - name
    - resource
    type: object
...
  handler.PermissionResponse:
    properties:
      action:
        description: 操作类型
        example: create
        type: string
      created_at:
        description: 创建时间
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        description: 权限描述
        example: 允许创建用户
        type: string
      display_name:
        description: 显示名称
        example: 创建用户
        type: string
      id:
        description: 权限ID
        example: 1
        type: integer
      name:
        description: 权限名称
        example: user:create
        type: string
      resource:
        description: 资源类型
        example: user
        type: string
      status:
        description: 状态：1-启用，0-禁用
        example: 1
        type: integer
      updated_at:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  handler.UpdatePermissionRequest:
    properties:
      description:
        description: 权限描述（可选）
        example: 允许创建用户
        type: string
      display_name:
        description: 显示名称（可选）
        example: 创建用户
        type: string
      status:
        description: 状态：1-启用，0-禁用（可选）
        example: 1
        type: integer
    type: object
...
  util.Response:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据（可选）
      error:
        description: 错误详情（可选）
        type: string
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个 Go 测试项目的 RESTful API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Test API
  version: "1.0"
paths:
  /permissions:
    get:
      consumes:
      - application/json
      description: 分页获取权限列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.PermissionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取权限列表
      tags:
      - 权限管理
    post:
      consumes:
      - application/json
      description: 创建一个新权限
      parameters:
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 创建权限
      tags:
      - 权限管理
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: 根据权限ID删除权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 删除权限
      tags:
      - 权限管理
    get:
      consumes:
      - application/json
      description: 根据权限ID获取权限详细信息
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取权限详情
      tags:
      - 权限管理
    put:
      consumes:
      - application/json
      description: 根据权限ID更新权限信息（支持部分更新）
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 更新权限
      tags:
      - 权限管理
...
schemes:
- http
- https
swagger: "2.0"